+++ database_tools.py
@@ -1,0 +1,16 @@
+import os
+
+import redis
+
+_database = None
+
+
+def get_database_connection():
+    global _database
+
+    if _database is None:
+        redis_bd_credentials = os.getenv('REDIS_BD_CREDENTIALS')
+        _database = redis.from_url(redis_bd_credentials)
+        _database.ping()
+
+    return _database
